+++ deepctr/__init__.py
@@ -1,4 +1,4 @@
 from .utils import check_version
 
-__version__ = '0.9.3'
+__version__ = '0.9.4'
 check_version(__version__)