+++ deepctr/models/sequence/dien.py
@@ -8,8 +8,8 @@ Reference:
 """
 
 import tensorflow as tf
-from tensorflow.python.keras.models import Model
-from tensorflow.python.keras.layers import (Concatenate, Dense, Permute, multiply, Flatten)
+from tensorflow.keras.models import Model
+from tensorflow.keras.layers import (Concatenate, Dense, Permute, multiply, Flatten)
 
 from ...feature_column import SparseFeat, VarLenSparseFeat, DenseFeat, build_input_features
 from ...inputs import get_varlen_pooling_list, create_embedding_matrix, embedding_lookup, varlen_embedding_lookup, \